+++ src/main/java/PodoeMarket/podoemarket/profile/dto/response/WorkListResponseDTO.java
@@ -35,9 +35,9 @@ public class WorkListResponseDTO {
             private String title;
             private String imagePath;
             private Boolean script;
-            private Integer scriptPrice;
+            private Long scriptPrice;
             private Boolean performance;
-            private Integer performancePrice;
+            private Long performancePrice;
             private ProductStatus checked;
         }
     }
